⚡️: my-page/info 페이지 - 컴포넌트 분리 및 코드 리펙토링
